Move authentication logic to extension method
